+++ glfw/wl_platform.h
@@ -292,7 +292,7 @@ typedef struct _GLFWlibraryWayland
 
     struct wl_surface*          cursorSurface;
     GLFWCursorShape             cursorPreviousShape;
-    uint32_t                    serial, input_serial;
+    uint32_t                    serial, input_serial, pointer_serial, keyboard_enter_serial;
 
     int32_t                     keyboardRepeatRate;
     monotonic_t                 keyboardRepeatDelay;